Chore: 폴더 구조민 코드 분리 작업 (로직 변화 없음)
