+++ users/urls.py
@@ -1,13 +1,12 @@
 from django.urls import path
-from users.views import (
-    SignUpAPIView,
-    SendVerificationCodeAPIVIew,
-    CheckVerificationCodeAPIView,
-    LoginAPIView,
-    LogoutAPIView,
-    CustomTokenRefreshView,
-    ChangePasswordAPIView
+from users.views.email_verification import (
+    SendVerificationCodeAPIVIew, CheckVerificationCodeAPIView
 )
+from users.views.signup import SignUpAPIView
+from users.views.auth.login import LoginAPIView
+from users.views.auth.logout import LogoutAPIView
+from users.views.auth.token import CustomTokenRefreshView
+from users.views.account import ChangePasswordAPIView
 
 urlpatterns = [
     path("send-code", SendVerificationCodeAPIVIew.as_view(), name="verification-email"),
@@ -17,5 +16,4 @@ urlpatterns = [
     path("logout", LogoutAPIView.as_view(), name="logout-user"),
     path("reissue-token", CustomTokenRefreshView.as_view(), name="reissue-token"),
     path("change-pwd", ChangePasswordAPIView.as_view(), name="change-pwd"),
-    # path("{user_id}/change-pwd", ChangePasswordAPIView.as_view(), name="change-pwd"),
 ]
